moved default group services from services view model into external presenter object  #309
